+++ public.ssl/modules/actes/class/ActesTransaction.class.php
@@ -1101,7 +1101,7 @@ class ActesTransaction extends DataObject {
 			return false;
  	  	}
 
- 	  	if ( ($this->files["acte"]['mimetype'] == 'application/xml') && (! in_array($ext,array('pdf')))){
+ 	  	if (  (isset($this->files["acte"]['mimetype']) && $this->files["acte"]['mimetype'] == 'application/xml') && (! in_array($ext,array('pdf')))){
  	  		$this->errorMsg = "Les pièces jointes doivent être au format PDF avec un acte au format XML";
  	  		return false;
  	  	}