+++ public.ssl/javascript/tedetis.js
@@ -57,7 +57,7 @@ function afficheWarning(){
       if(liste[i].checked) n++;
     }
   }
-  var msg = "Voulez-vous vraiment affecter les " + n + " transactions selectionnees ? ";
-  msg += "Cette action est non reversible et est sous votre entiere responsabilite.";
+  var msg = "Voulez-vous vraiment affecter les " + n + " transactions selectionnées ? ";
+  msg += "Cette action n'est pas réversible et est sous votre entière responsabilité.";
   return confirm(msg);
 }