+++ public.ssl/modules/helios/index.php
@@ -153,7 +153,7 @@ function afficheWarning(){
       if(liste[i].checked) n++;
     }
   }
-  var msg = "Voulez-vous vraiment affecter les " + n + " transactions sélectionnés ? ";
+  var msg = "Voulez-vous vraiment affecter les " + n + " transactions sélectionnées ? ";
   msg += "Cette action est non réversible et est sous votre entière responsabilité";
   return confirm(msg);
 }